requirements/commom.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6